fix(settings): restore bookmarklets subpage
Promote the existing bookmarklet generator to a first-class
/settings/bookmarklets subpage so it is discoverable again.

- Extract the bookmarklet panel out of skills-section into a reusable
  bookmarklets-section component.
- Register the visible Bookmarklets route in the Settings registry while
  keeping the /settings/skills?skill=__bm deep link working.
- Update registry, route, and settings tests plus the redesign spec.

The generator's /new?skill=&ref=&auto=&key= contract, launch-key handling,
and same-origin fetch are unchanged.
